refactor(pr10): rehome resolveChangeValue into BaseActiveEffect/logic/ (G5i.6-G5i.7)
Moves src/entities/activeEffects/BaseActiveEffect/resolveChangeValue.mts into a new logic/ subfolder with a barrel index, and repoints both direct importers (ItemDnd35e, ActorDnd35e) plus the two unit-test imports.
