+++ src/entities/activeEffects/BaseActiveEffect/logic/index.mts
@@ -1,0 +1,6 @@
+export {
+  getEffectContexts,
+  resolveActiveEffectChange,
+  resolveActiveEffectChangeValue,
+  resolveMaskedActiveEffectChangeValue,
+} from './resolveChangeValue.mjs';
